+++ src/features/speedrun/speedrunTimer.ts
@@ -1,7 +1,14 @@
-import { RENDER_FRAMES_PER_SECOND } from "isaacscript-common";
+import {
+  getScreenBottomY,
+  RENDER_FRAMES_PER_SECOND,
+  RESOLUTION_FULL_SCREEN,
+} from "isaacscript-common";
+import { RACE_TIMER_POSITION_X, RACE_TIMER_POSITION_Y } from "../../constants";
 import { TimerType } from "../../enums/TimerType";
+import { config } from "../../modConfigMenu";
 import * as timer from "../../timer";
 import { shouldDrawRaceTimer } from "../race/raceTimer";
+import { isOnFirstCharacter } from "./speedrun";
 import { v } from "./v";
 
 // ModCallback.POST_RENDER (2)
@@ -13,6 +20,7 @@ export function postRender(): void {
   }
 
   drawSpeedrunTimer();
+  drawSpeedrunCharacterTimer();
 }
 
 function drawSpeedrunTimer() {
@@ -30,4 +38,42 @@ function drawSpeedrunTimer() {
   const seconds = elapsedFrames / RENDER_FRAMES_PER_SECOND;
 
   timer.draw(TimerType.RACE_OR_SPEEDRUN, seconds);
+}
+
+function drawSpeedrunCharacterTimer() {
+  if (!config.characterTimer) {
+    return;
+  }
+
+  // If we are on the first character, the two timers would be identical.
+  if (isOnFirstCharacter()) {
+    return;
+  }
+
+  // If the resolution is too tiny, the second timer would overlap with the trinket UI.
+  const screenBottomY = getScreenBottomY();
+  if (screenBottomY < RESOLUTION_FULL_SCREEN.Y) {
+    return;
+  }
+
+  const renderFrameCount = Isaac.GetFrameCount();
+
+  // Find out how much time has passed since the last "split" (e.g. when the last checkpoint was
+  // touched).
+  let elapsedFrames: int;
+  if (v.run.finished && v.run.finishedFrames !== null) {
+    elapsedFrames = v.run.finishedFrames;
+  } else if (v.persistent.startedCharacterFrame === null) {
+    elapsedFrames = 0;
+  } else {
+    elapsedFrames = renderFrameCount - v.persistent.startedCharacterFrame;
+  }
+  const seconds = elapsedFrames / RENDER_FRAMES_PER_SECOND;
+
+  timer.draw(
+    TimerType.SPEEDRUN_CHARACTER,
+    seconds,
+    RACE_TIMER_POSITION_X,
+    RACE_TIMER_POSITION_Y + 15,
+  );
 }